Started the creature building section.
Added nice formatting to the creature enums.
Split the creature fields that could be useful in encounter building
from those that are just nice to have.
Made a creature building window the same as the spell builder.
 - Finished its front end with the barest set of fields.
Added a creatureBuilder spawner to the main window.

I created a copy of the old creature list.
Also added the PHB's cleric cantrips.