Added the rest of the information to spells. *WARNING* save wont work.
Overhauling spells so that the spell system (and doubly so the spellbook
system) could be useful by itself, and useful for players.

Added archetypes to the classes enum.
Moved classes enum out of spells, and areas enum out of actions.

Added all of the useful information getters to the spell interface.
 - Updated the toXML method for the new info.
 - - Updated the parser to read most of the new XML.

Copied the spell list so I don't lose it with the new format.