Set up filters and most of the known list for the spellbook.
Updated spellbook filters to match the new spell builder ones.
Modified the filters in the spellbook window to also change a query on
the known list.
Set up buttons for adding spells to the known list.
Set up the frame for the known list, and started work on adding and
removing entries.
 - Need to purge and refill the list on update for spacing.

Changed a little of the spell format around the effects.